ptrdiff_t in 64-bit may not be convertible to size_t or unsigned
git-svn-id: svn+ssh://svn.corp.imvu.com/home/svnrepos/trunk/stable/Source/cal3d@111750 07c76cb3-cb09-0410-85de-c24d39f1912e

+++ src/cal3d/saver.cpp
@@ -651,7 +651,7 @@ bool CalSaver::saveCoreSubmesh(std::ostream& os, CalCoreSubmesh* pCoreSubmesh) {
         }
         ++nextVertex;
 
-        if (!CalPlatform::writeInteger(os, nextVertex - currentInfluence)) {
+        if (!CalPlatform::writeInteger(os, size_t(nextVertex - currentInfluence))) {
             CalError::setLastError(CalError::FILE_WRITING_FAILED, __FILE__, __LINE__, "");
             return false;
         }
@@ -1190,7 +1190,7 @@ bool CalSaver::saveXmlCoreMesh(const std::string& strFilename, CalCoreMesh* pCor
 
             TiXmlElement vertex("VERTEX");
             vertex.SetAttribute("ID", vertexId);
-            vertex.SetAttribute("NUMINFLUENCES", nextVertex - currentInfluence);
+            vertex.SetAttribute("NUMINFLUENCES", size_t(nextVertex - currentInfluence));
 
             // write the vertex data
